rearrange the modules
